+++ src/main/java/com/ddudu/application/port/in/auth/TokenRefreshUseCase.java
@@ -1,4 +1,4 @@
-package com.ddudu.application.port.in;
+package com.ddudu.application.port.in.auth;
 
 import com.ddudu.application.domain.authentication.dto.request.TokenRefreshRequest;
 import com.ddudu.application.domain.authentication.dto.response.TokenResponse;
