+++ src/main/java/com/ddudu/application/service/goal/CreateGoalService.java
@@ -8,8 +8,8 @@ import com.ddudu.application.domain.goal.exception.GoalErrorCode;
 import com.ddudu.application.domain.goal.service.GoalDomainService;
 import com.ddudu.application.domain.user.domain.User;
 import com.ddudu.application.port.in.goal.CreateGoalUseCase;
-import com.ddudu.application.port.out.UserLoaderPort;
 import com.ddudu.application.port.out.goal.SaveGoalPort;
+import com.ddudu.application.port.out.user.UserLoaderPort;
 import jakarta.transaction.Transactional;
 import java.util.MissingResourceException;
 import lombok.RequiredArgsConstructor;
