+++ src/main/java/com/ddudu/application/service/goal/RetrieveGoalService.java
@@ -6,7 +6,7 @@ import com.ddudu.application.domain.goal.dto.response.GoalResponse;
 import com.ddudu.application.domain.goal.exception.GoalErrorCode;
 import com.ddudu.application.port.in.goal.RetrieveGoalUseCase;
 import com.ddudu.application.port.out.goal.GoalLoaderPort;
-import com.ddudu.application.port.out.UserLoaderPort;
+import com.ddudu.application.port.out.user.UserLoaderPort;
 import java.util.MissingResourceException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.transaction.annotation.Transactional;
