+++ src/main/java/com/ddudu/infrastructure/api/adapter/ExternalApiAdapter.java
@@ -2,7 +2,7 @@ package com.ddudu.infrastructure.api.adapter;
 
 import com.ddudu.application.domain.authentication.dto.request.SocialRequest;
 import com.ddudu.application.domain.user.domain.vo.AuthProvider;
-import com.ddudu.application.port.out.SocialResourcePort;
+import com.ddudu.application.port.out.auth.SocialResourcePort;
 import com.ddudu.infrastructure.annotation.DrivenAdapter;
 import com.ddudu.infrastructure.api.handler.SocialResourceHandler;
 import com.ddudu.infrastructure.api.response.SocialResource;
