+++ src/main/java/com/ddudu/presentation/api/controller/AuthController.java
@@ -4,8 +4,8 @@ import com.ddudu.application.domain.authentication.dto.request.SocialRequest;
 import com.ddudu.application.domain.authentication.dto.request.TokenRefreshRequest;
 import com.ddudu.application.domain.authentication.dto.response.TokenResponse;
 import com.ddudu.application.domain.user.domain.enums.ProviderType;
-import com.ddudu.application.port.in.SocialLoginUseCase;
-import com.ddudu.application.port.in.TokenRefreshUseCase;
+import com.ddudu.application.port.in.auth.SocialLoginUseCase;
+import com.ddudu.application.port.in.auth.TokenRefreshUseCase;
 import com.ddudu.old.auth.dto.response.MeResponse;
 import com.ddudu.old.auth.service.AuthService;
 import com.ddudu.presentation.api.annotation.Login;
