+++ src/test/java/com/ddudu/application/service/CreateGoalServiceTest.java
@@ -11,10 +11,10 @@ import com.ddudu.application.domain.goal.dto.response.GoalIdResponse;
 import com.ddudu.application.domain.goal.exception.GoalErrorCode;
 import com.ddudu.application.domain.goal.service.GoalDomainService;
 import com.ddudu.application.domain.user.domain.User;
+import com.ddudu.application.port.out.auth.SignUpPort;
 import com.ddudu.application.port.out.goal.GoalLoaderPort;
 import com.ddudu.application.port.out.goal.SaveGoalPort;
-import com.ddudu.application.port.out.SignUpPort;
-import com.ddudu.application.port.out.UserLoaderPort;
+import com.ddudu.application.port.out.user.UserLoaderPort;
 import com.ddudu.application.service.goal.CreateGoalService;
 import com.ddudu.fixture.BaseFixture;
 import com.ddudu.fixture.GoalFixture;
