+++ src/test/java/com/ddudu/application/service/RetrieveAllGoalsServiceTest.java
@@ -7,10 +7,10 @@ import com.ddudu.application.domain.goal.domain.Goal;
 import com.ddudu.application.domain.goal.dto.response.GoalSummaryResponse;
 import com.ddudu.application.domain.goal.exception.GoalErrorCode;
 import com.ddudu.application.domain.user.domain.User;
+import com.ddudu.application.port.out.auth.SignUpPort;
 import com.ddudu.application.port.out.goal.GoalLoaderPort;
 import com.ddudu.application.port.out.goal.SaveGoalPort;
-import com.ddudu.application.port.out.SignUpPort;
-import com.ddudu.application.port.out.UserLoaderPort;
+import com.ddudu.application.port.out.user.UserLoaderPort;
 import com.ddudu.application.service.goal.RetrieveAllGoalsService;
 import com.ddudu.fixture.BaseFixture;
 import com.ddudu.fixture.GoalFixture;
