+++ src/test/java/com/ddudu/application/service/SocialLoginServiceTest.java
@@ -9,9 +9,10 @@ import com.ddudu.application.domain.authentication.dto.response.TokenResponse;
 import com.ddudu.application.domain.user.domain.User;
 import com.ddudu.application.domain.user.domain.enums.ProviderType;
 import com.ddudu.application.domain.user.domain.vo.AuthProvider;
-import com.ddudu.application.port.out.SignUpPort;
-import com.ddudu.application.port.out.SocialResourcePort;
-import com.ddudu.application.port.out.UserLoaderPort;
+import com.ddudu.application.port.out.auth.SignUpPort;
+import com.ddudu.application.port.out.auth.SocialResourcePort;
+import com.ddudu.application.port.out.user.UserLoaderPort;
+import com.ddudu.application.service.auth.SocialLoginService;
 import com.ddudu.fixture.UserFixture;
 import jakarta.transaction.Transactional;
 import java.util.Optional;
