+++ src/test/java/com/ddudu/application/service/TokenRefreshServiceTest.java
@@ -11,10 +11,11 @@ import com.ddudu.application.domain.authentication.dto.response.TokenResponse;
 import com.ddudu.application.domain.authentication.exception.AuthErrorCode;
 import com.ddudu.application.domain.authentication.service.AuthDomainService;
 import com.ddudu.application.domain.user.domain.User;
-import com.ddudu.application.port.out.SignUpPort;
-import com.ddudu.application.port.out.TokenLoaderPort;
-import com.ddudu.application.port.out.TokenManipulationPort;
-import com.ddudu.application.port.out.UserLoaderPort;
+import com.ddudu.application.port.out.auth.SignUpPort;
+import com.ddudu.application.port.out.auth.TokenLoaderPort;
+import com.ddudu.application.port.out.auth.TokenManipulationPort;
+import com.ddudu.application.port.out.user.UserLoaderPort;
+import com.ddudu.application.service.auth.TokenRefreshService;
 import com.ddudu.fixture.UserFixture;
 import jakarta.transaction.Transactional;
 import java.util.List;
